feat: embedding upgrades, HDC-LTC improvements, and primitive fixes
Embedding System:
- Add BGE-M3 backend (bge_m3.rs) as modern replacement for BERT
- Add LayerExtractor trait for unified layer access across models
- Enhance multi_model_extractor with configurable backends
- Update bert_layer_extractor with improved error handling

HDC-LTC Architecture:
- Add hd_ltc_codec.rs for HDC-LTC state encoding/decoding
- Improve hdc_ltc_neuron.rs with better initialization and dynamics
- Wire new components into hdc/mod.rs

Primitive System:
- Fix HV16::bundle() API calls in primitive_discovery.rs
- Fix double-reference iteration patterns
- Update primitive_consciousness.rs bindings

Language/LLM:
- Add llm_organ.rs unified LLM interface
- Enhance llm_backend.rs with better provider abstraction
- Improve emotional_core.rs sentiment processing

Perception:
- Enhance audio.rs with better feature extraction
- Add hierarchical CfC support in dynamics/mod.rs

+++ symthaea-hlb/src/consciousness/phi_architecture_search.rs
@@ -1444,7 +1444,7 @@ mod tests {
         // At least some parameter should have changed
         // (with mutation rate 1.0, all parameters mutate)
         // Use epsilon-based comparison for floating-point values
-        const EPSILON: f64 = 1e-9;
+        const EPSILON: f32 = 1e-9;
         assert!(
             (genome.connection_density - original_density).abs() > EPSILON ||
             genome.num_nodes != 16 ||
